+++ src/main/java/com/chaewookim/accountbookformoms/domain/transaction/dto/request/transaction/TransactionFixRequest.java
@@ -1,4 +1,4 @@
-package com.chaewookim.accountbookformoms.domain.transaction.dto.request;
+package com.chaewookim.accountbookformoms.domain.transaction.dto.request.transaction;
 
 import jakarta.validation.constraints.NotNull;
