+++ src/main/java/com/chaewookim/accountbookformoms/domain/transaction/dto/request/transaction/TransactionMemoRequest.java
@@ -1,4 +1,4 @@
-package com.chaewookim.accountbookformoms.domain.transaction.api;
+package com.chaewookim.accountbookformoms.domain.transaction.dto.request.transaction;
 
 import jakarta.validation.constraints.NotBlank;
