+++ src/main/java/com/chaewookim/accountbookformoms/domain/transaction/dto/request/transaction/TransactionUpdateRequest.java
@@ -1,6 +1,5 @@
-package com.chaewookim.accountbookformoms.domain.transaction.dto.response;
+package com.chaewookim.accountbookformoms.domain.transaction.dto.request.transaction;
 
-import com.chaewookim.accountbookformoms.domain.transaction.entity.Transaction;
 import com.chaewookim.accountbookformoms.domain.transaction.enums.TransactionTypeEnum;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
@@ -9,7 +8,7 @@ import jakarta.validation.constraints.Size;
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
-public record TransactionResponse(
+public record TransactionUpdateRequest(
         @NotNull
         Long id,
 
@@ -40,18 +39,4 @@ public record TransactionResponse(
 
         LocalDateTime repeatDate
 ) {
-    public static TransactionResponse from(Transaction transaction) {
-        return new TransactionResponse(
-                transaction.getId(),
-                transaction.getTitle(),
-                transaction.getMemo(),
-                transaction.getUserId(),
-                transaction.getAssetId(),
-                transaction.getCategoryId(),
-                transaction.getAmount(),
-                transaction.getType(),
-                transaction.getIsFixed(),
-                transaction.getRepeatDate()
-        );
-    }
 }
